feat: Add market-implied betting probability features
- Replace inverse odds with more advanced probability features
- Add 11 new horse-specific probability features derived from betting odds:
  - prob_win: single-horse win probability
  - prob_place: single-horse place probability
  - prob_frame_quinella: probability for horse's frame in bracket quinella
  - prob_quinella: probability of horse appearing in quinella
  - prob_quinella_place: probability of horse in quinella place
  - prob_exacta_first: probability of horse winning in exacta
  - prob_exacta: probability of horse appearing in exacta
  - prob_trio: probability of horse appearing in trio
  - prob_trifecta_first: probability of horse winning in trifecta
  - prob_trifecta_first_second: probability of horse in 1st or 2nd in trifecta
  - prob_trifecta: probability of horse appearing anywhere in trifecta
- Update config to use new probability features in numerical_features
- Add dtype=np.float32 to improve memory usage in odds arrays

+++ horguesc/core/betting/odds_loader.py
@@ -148,8 +148,8 @@ class OddsLoader:
         race_id_to_index = {race_id: i for i, race_id in enumerate(race_ids)}
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), self.max_horse_number), np.nan)
-        
+        result = np.full((len(race_ids), self.max_horse_number), np.nan, dtype=np.float32)
+
         # Build the query to get the latest odds for each race
         query = f"""
             WITH latest_odds AS (
@@ -209,8 +209,8 @@ class OddsLoader:
         race_id_to_index = {race_id: i for i, race_id in enumerate(race_ids)}
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), self.max_horse_number), np.nan)
-        
+        result = np.full((len(race_ids), self.max_horse_number), np.nan, dtype=np.float32)
+
         # Build the query to get the latest odds for each race
         query = f"""
             WITH latest_odds AS (
@@ -273,7 +273,7 @@ class OddsLoader:
         num_combinations = self.max_frame_number * (self.max_frame_number + 1) // 2
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
@@ -353,7 +353,7 @@ class OddsLoader:
         num_combinations = self.max_horse_number * (self.max_horse_number - 1) // 2
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
@@ -433,7 +433,7 @@ class OddsLoader:
         num_combinations = self.max_horse_number * (self.max_horse_number - 1) // 2
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
@@ -513,7 +513,7 @@ class OddsLoader:
         num_combinations = self.max_horse_number * (self.max_horse_number - 1)
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
@@ -593,7 +593,7 @@ class OddsLoader:
         num_combinations = self.max_horse_number * (self.max_horse_number - 1) * (self.max_horse_number - 2) // 6
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
@@ -681,7 +681,7 @@ class OddsLoader:
         num_combinations = self.max_horse_number * (self.max_horse_number - 1) * (self.max_horse_number - 2)
         
         # Initialize the result array with NaN
-        result = np.full((len(race_ids), num_combinations), np.nan)
+        result = np.full((len(race_ids), num_combinations), np.nan, dtype=np.float32)
         
         # Build the query to get the latest odds for each race
         query = f"""
